+++ CHANGES.md
@@ -1,3 +1,14 @@
+## 0.0.53
+
+Connectors:
+
+-   Allowed Ckan connector to pull datasets belongs to a specified organisation
+-   Added `presetRecordAspects` & `extra` parameters supports to all connectors
+
+Dataset Page:
+
+-   When `dcat-dataset-strings.creation.isOpenData` is set, show `Public / Private` label accordingly
+
 ## 0.0.52
 
 Authentication:
@@ -73,6 +84,7 @@ Ops:
 -   Added `cloudsql-db-credentials` to create-secrets tool
 -   Only runtime dependencies will be included by docker image build script
 -   Fixed an file selector error in create-secrets script when current directory & non of its sub directory has \*.json file
+-   Changed SQL for querying aspects by path/value so that it could take advantage of JSONB indices.
 
 Security: